moved JsDoc components into their own directory
