+++ src/rsg-components/JsDoc/Arguments/index.js
@@ -1,0 +1,1 @@
+export { default } from './ArgumentsRenderer';
