+++ src/rsg-components/JsDoc/Authors/index.js
@@ -1,0 +1,2 @@
+export { default } from './AuthorsRenderer';
+
