+++ src/rsg-components/JsDoc/Deprecated/index.js
@@ -1,0 +1,1 @@
+export { default } from './DeprecatedRenderer';
