+++ src/rsg-components/JsDoc/Links/index.js
@@ -1,0 +1,1 @@
+export { default } from './LinksRenderer';
