+++ src/rsg-components/JsDoc/Since/index.js
@@ -1,0 +1,1 @@
+export { default } from './SinceRenderer';
