+++ src/rsg-components/JsDoc/Version/index.js
@@ -1,0 +1,2 @@
+export { default } from './VersionRenderer';
+
