--- src/rsg-components/JsDoc/index.js
@@ -1,6 +1,0 @@
-export { default as JsDocArguments } from './ArgumentsRenderer';
-export { default as JsDocAuthors } from './AuthorsRenderer';
-export { default as JsDocDeprecated } from './DeprecatedRenderer';
-export { default as JsDocLinks } from './LinksRenderer';
-export { default as JsDocSince } from './SinceRenderer';
-export { default as JsDocVersion } from './VersionRenderer';
